feat: Add equipment management features
- Updated EquipmentItem and EquipmentCategory models to include an ID field.
- Modified Exercise model to include equipmentId for associating exercises with equipment.
- Refactored EquipmentsScreen to use ListCategories and SelectedCategoryList for better UI.
- Implemented ChooseEquipmentScreen for selecting equipment with search functionality.
- Enhanced EditWorkoutsScreen to allow adding equipment to exercises.
- Created admin_display_categories and admin_list_category_widget for displaying categories and equipment.
- Updated models.dart to export new equipment models.
