+++ lib/models.dart
@@ -1,0 +1,5 @@
+export 'package:academia_unifor/models/notifications.dart';
+export 'package:academia_unifor/models/exercise.dart';
+export 'package:academia_unifor/models/equipment.dart';
+export 'package:academia_unifor/models/users.dart';
+export 'package:academia_unifor/models/workout.dart';
