+++ lib/widgets.dart
@@ -8,3 +8,6 @@ export 'package:academia_unifor/widgets/contact_widget.dart';
 
 // admin
 export 'package:academia_unifor/widgets/admin/admin_convex_bottom_bar.dart';
+export 'package:academia_unifor/widgets/admin/admin_display_categories.dart';
+export 'package:academia_unifor/widgets/admin/admin_list_category_widget.dart';
+export 'package:academia_unifor/widgets/admin/admin_list_equipment_widget.dart';
